Add the tiles to git
Add the tile image and data file to git.
The .json will have a special object called `variants`
which lists the different types of tile and the amounts
of variants for each as well as their different chances for
appearing. The chance will be an x tickets of all the tickets
kind of chance.